chore(wizard-refactor): extract _shared/ensure-institution-and-domain
Depends on resolvers/institution-by-name + resolvers/infer-type-from-name
(both extracted earlier). Drops monolith's direct use of inferTypeFromName.
Dispatcher 11/11 + lint 17/17 green.
